+++ frontend/app/src/shared/components/layout/breadcrumb-navigation/breadcrumb-account-profile.tsx
@@ -1,7 +1,7 @@
 import { constructPath } from "@/shared/api/rest/fetch";
 import { BreadcrumbItem, Breadcrumbs } from "@/shared/components/aria/breadcrumbs";
 
-export function BreadcrumbProfile() {
+export function BreadcrumbAccountProfile() {
   return (
     <Breadcrumbs data-testid="breadcrumb-profile">
       <BreadcrumbItem href={constructPath("/profile")}>Account settings</BreadcrumbItem>
